Add series path saving and loading
Up until now, the only way to watch a series was to specify a direct path to it. Now, you can play a series simply by typing in a set name for it.
For example, to save a series: "./anitrack toradora -p ~/anime/Toradora". And to play it later: "./anitrack toradora"
You can also play a series just from its path like before, like so: "./anitrack -p ~/anime/Toradora"