Add functionality to update the AniList access token if invalid
If the application's access to a user's AniList account is revoked, the user would have to delete the application's configuration file in order to set a new access token. Now, the invalid token will be detected at launch and the URL to authorize the application will be opened to allow access again.