Remove --dontsavetoken command line argument
Not saving the auth token will make you have to authorize the application every time it's launched, which will essentially spam your list of authorized applications on AniList.

This command line argument was initially added when MyAnimeList was the only backend supported, which meaned that the application had to store the account password, but now that those days are over and the only thing that needs to be stored is an OAuth token, it does not seem beneficial to have this option anymore.