[core] Fix some types of custom episode matchers never detecting episodes
If you tried to use a custom regex pattern on a series that had underscores in the filename, it would never match because the underscores were being replaced with spaces in episode parsing to aid in matching with the default matcher.
This led to confusing behavior where you had to use spaces instead of underscores in the custom matcher for it to work.

This change explicitly matches for special symbols like _ and . in the default matcher and only replaces underscores when extracting the name of the series to aid in getting good search results on AniList.