[series] Refactor the way season states are handled (again)
The new way of handling state goes somewhat back to the old method of state management, which involved cloning.
This was originally changed to try to avoid potential bugs in the future from doing something like saving an out of date state due to having instances of the data in multiple places.

The new method involves working on a direct copy of the season state data in the Series struct, and updating the save file instance of it automatically whenever we want to write the file to disk.

The primary reason for refactoring this again was to reduce the sheer amount of array acceses needed simply to update / read the state data, which could have also introduced bugs in the future with things like an out of bounds panic.