[series] Add functionality to select from multiple found series in a directory
This commit completely overhauls the existing method for discovering episode files.

Before, any detected title of a file that was unique would completely prevent you from continuing until you either deleted the conflicting files or moved them.
Now, when multiple titles are detected, the program will allow you to choose which set you want to use. This avoids conflicts with things like special episodes, which are often thrown in the same directory as the main series.

This addition does not allow you to watch a separately identified series that's in the same directory, but it should alleviate some pain when trying to watch some series' for the first time.