[series] Add subseries functionality
This commit introduces subseries, which allows you to watch multiple series within the same folder.

Previously, you would only be allowed to select one series in a folder if multiple were detected, which made it annoying to watch specials or movies that were downloaded to the same place.

Now, simply launching the program with the name of the subseries you want to watch will allow you to select any series found in a folder and have it be treated as if it was in its own folder.