[anup] Remove functionality to play episodes in a loop
This feature does not seem appropriate to include, given that single episodes can be played.

It would be better to leave it to users to create a shell script to run the program in a loop instead, as they would be able to make it
fit their needs much better than built-in functionality can offer.