[anup] Add a functional terminal user interface
This commit adds a TUI that can be accessed via the -i flag. It currently supports navigating through different seasons of a series with
the up and down arrow keys, and playing the next unwatched episode of the selected season with enter. Pressing q will exit the interface.

It will also show you the amount of time needed to progress the series in real-time whenever you start watching an episode, as well as
when the series is unsynced with AniList.

Another change is that now series info is no longer lazily fetched from AniList. This removes the chance of being unable to watch a
different season of a series that wasn't prefetched while offline. This change should have been split into a different commit, but it
was made deep enough into implementing the TUI that it wouldn't be feasible to cherry-pick it.