[anup] Add functionality to select different series in the TUI
This change introduces the ability to scroll through every series that has been added to the program in the same way that you could
already scroll through different seasons. Two new keybindings has been added to accomodate this change: the left & right arrow keys will
now switch between series and season selection, with a '>' appearing in the list that is active. Finally, the last watched series will
be updated if you watch a series other than the one selected when you launched the program.