[anime] Don't clean parsed episode titles
This was implemented when the program was written to only work with MyAnimeList. At that time, cleaning episode files was necessary
because its search function struggled to get results when the search term contained unrelated characters. Now, it actually reduces the
accuracy of search results when used with AniList.

Any remote service implemented in the future could easily implement its own cleaning of search terms if necessary anyway.