[anup] Don't sync the last watched season when launching the TUI
This is being removed to reduce possible ambiguity. Right now, none of the other seasons of the series are synced from the remote at
launch, and since there is no feasible way to indicate to the user that a season is actually synced to the remote entry for it (as it can
change at any time), it can lead to somewhat confusing behavior as to what is and isn't up to date.

It's better to let the user manually sync the series from the remote service via the r key instead, or simply let it get synced when the
user watches the series.