[anup] Add command system to TUI to replace most keybindings
This adds a fully-integrated command system with suggestions to allow for much more powerful
things to be done within the UI, and reduce the mental burden of trying to remember a lot of keybindings.

Command suggestions are based off the current input, and can be completed by pressing tab. Currently, suggestions
are only provided for the command name itself and not its arguments, but their usage will eventually be documented
in the readme, and may eventually be added as part of the suggestions.

The only lingering issue is that the cursor is not visible while entering a command. We cannot set the cursor
position while drawing due to borrow issues, and simply making it visible makes it jump around because of the way
suggestion text is implemented, so there doesn't appear to be an easy way to implement this.

The command prompt can be entered by pressing ':'.