refactor codebase to not explicitly handle merged seasons within a series
Tracking merged seasons adds a lot of complexity to the entire codebase (as shown by how many places this commit touches),
and it's not even a commonly utilized feature.

A new CLI tool will be developed instead to make changes to the episode files themselves, as the feature is quite
useful in the rare circumstances where you do actually need it.

This refactor also removes the concept of having separate files for every different piece of data, as that also added
a good amount of complexity to everything.