anup: refactor codebase to use SQLite for data storage
This changes the data storage scheme from MessagePack compressed files + a TOML name map file to SQLite. The main reason for this change is that it
allows for much easier debugging of stored data.

There are two main drawbacks to this change:

1. The first run of the program now takes 12+ ms to create the database.
2. Fetching series information is ~2x slower (although the overall time is negligible).

Saving a series to the database tends to be slightly faster once things are cached, even with the naive way it's currently implemented.