anime: return Iterator from RemoteService::search_info_by_name
This allows us to lazily search for the series info we want, so we don't waste time converting between types we may not need.
Although it does involve an allocation for the Box, our current use case in anup and anisplit makes it a net benefit as both programs only want one of the results.