treewide: replace rusqlite with diesel
This changes a lot of little things all over the place to fit in well with diesel. Existing series databases will still work, although the table names have changed so existing series will not be shown. Therefore, it's recommended to either migrate the database manually to the new table names or simply delete it and add the series back.