refactor codebase to support series info selection in anup
This adds support for selecting from a list of found series in anup if it can't confidently select it on its own.

This whole commit ended up being much larger than expected, as it includes a complete refactor of the detect library and minor refactors in the TUI module to accomodate this feature. Some of these refactors could have been split into multiple commits, but I was too lazy to do it.