anup: refactor TUI to use a component system
This reduces the overall complexity of the TUI by using a new trait called Component to implement the tick / key / draw logic for each panel, as well as for meta items like playing the next unwatched series episode.

This system also makes use of a new enum called LogResult, which ensures we cannot accidently exit the TUI from error propagation, and makes each component easier to debug as the Err variant requires a description.