anup: try to lookup series info by the title of the best matching directory
The name of the directory a series is in is likely to be the full series title. This also reduces the chance of a false positive series match since we now only need to match our given nickname with the titles found in the series directory, rather than all possible results from the remote service.