anup: simplify series processing in the TUI
This changes the way errors are handled in the TUI by removing the panel that would normally show them, and instead now shows them in the status log. This allows us to remove the Invalid state from the SeriesStatus enum, which simplifies a lot of areas within the TUI.

A new struct called SeriesData that groups all of the information we need from the remote service is also introduced to simplify some constructors for the Series struct.