anime: Refactor episode parsing to use new default parser
This removes the EpisodeMatcher struct and replaces it with a new enum called EpisodeParser that allows for a default or custom parser (via a new EpisodeRegex struct) to be used.

The new default parser uses nom for parsing, instead of regex. This allows us to match many more formats without forcing the user to come up with custom regex patterns.