anup: Replace add command with a new dedicated panel
This adds a dedicated panel to add a new series to the program. It features realtime input validation and a preview of the detected series based off of current input. The new panel can be accessed by pressing the 'a' key.

Internally, this required a good chunk of changes to core data structures and introduces a new struct called SeriesPath for managing the absolute / relative path to a series.