anup: Add multi-user support via a user management panel
This adds a user management panel that can be accessed with the 'u' key. Its purpose is to make the onboarding process easier, and to make supporting new remote services easier in the future.

To support this feature, a new way of storing users has been introduced. All users are now stored at ~/.local/share/anup/users.mpack. The old ~/.config/anup/token.toml file should be deleted, as it is no longer used.