anup: Display series errors in the info panel
This replaces the Unloaded variant of the SeriesStatus enum with an Error one that stores the last encountered error for the series so it can be displayed when the series is selected.

This means that lazy series loading has been removed, but it should not really make a difference to performance as series loading has been pretty fast since episode parsing was revamped.