Implement categorized episode detection
This implements handling for episodes like specials and OVA's that are in the same folder as the main series.

Currently, anup will always select season episodes if multiple episode types are present.
If you want to watch the OVA's / ONA's / specials / movies of a series, you will need to use the anisplit tool to separate them into different folders.