anup: Add dedicated panel for series splitting
This adds a new panel that can be accessed with the 's' key that will find merged seasons / episode categories for the currently selected series and allow you split them up by pressing the 's' key again.
Once a series has been split up, you can select each of them with the arrow keys and press enter to easily add them to the program.

This makes the anisplit tool obsolete, and will be removed soon.