Use new custom format when specifying custom episode patterns to replace regex-based ones
Until now, the regex crate has been used to let users specify patterns to match episodes with when the default parser could not detect an episode format.
The problem with this is that there is a few "gotchas" that you need to be aware of and can be confusing if you aren't already familiar with regex in general.
A prime example of one of these issues is escaping brackets, which has to be done most of the time you specify a custom pattern as most filenames tend to have them.

To solve this (and save almost 1 MB from the release built & stripped binary), a new custom pattern syntax has been added. This new syntax matches filenames 1-to-1, unless * or # is specified.
The * character acts as a wildcard and matches everything in the filename until the next character in the pattern is found. The # character acts as an episode number marker and looks for numbers
in the filename where it is specified.

For example, the pattern "[*] series title -*- ep#" will detect 12 as the episode number in this filename: [Spaced Tag] Series Title - Episode Description - EP12.mkv