anup: Autofill name & path inputs when adding a series
When you add a series, the program will look for the last modified folder in the set series_dir and generate a nickname for it via the generate_nickname function added in 33e7e48b72a7dcae6a773a03a212976360a40cf0.
If a nickname could be generated, the name input will be set to it and the path input will also be set to the path of the last modified series.

The autofills are simply placeholders and are overridden when input is received. You can use the right arrow key to insert the placeholder into the input to modifiy it.