anup: Add edit mode to add series panel to replace set command
This removes the set command and implements an edit mode for the add series panel to replace it. The panel can be opened by pressing the 'e' key while the series you want to modify is selected.

Functionally, the panel works the same way as when adding a new series to the program, except that the inputs are prefilled with the current information for the series.