anup: Replace delete command with new panel
This replaces the delete command with a new dedicated panel that can be opened by pressing Shift + D.
Once in the panel, you can press D to toggle whether or not the episode files on disk will be deleted as well.

Once ready, you can then press enter to delete the series from the program and (if toggled) the episode files on disk.