anup: Implement proper key (de)serialization
This change may break existing config files. You can fix it by either removing the existing config or setting the play_next_episode field to "enter".

This commit allows keybindings to be specified more naturally in the config file and allows modifiers to be set.
As an example, you can now set the key to play episodes to something like "shift + a", "ctrl + shift", or "alt + f1".