anup: Optimize rendering time of main panel with tui-utils library
This uses the TextFragment widget and grid_pos() function from the https://github.com/Acizza/tui-utils library to reduce the rendering time of the main panel. On my machine, it brings the average rendering time down from 0.13 ms to 0.09 ms, which saves nearly ~60 ms of CPU time over the course of a 24 minute anime episode.