anup: Add independent draw method for each component & remove Draw trait from component module
The Draw trait wasn't really used for anything specific, and imposed limitations to components that implemented it (such as always requiring a mutable borrow).
All components now have their own specific draw method, which allows for most of them to only take immutable borrows and remove unused state parameters.

The Input and SeriesList components have also been refactored in this commit to draw immutably, which allowed for simplifications in the main tui module as well.